requirements.txt
yfinance==0.2.50
pandas==2.2.3
numpy==2.1.3
requests==2.32.3
feedparser==6.0.11
anthropic==0.39.0
python-dotenv==1.0.1
python-dateutil==2.9.0
lxml==5.3.0
beautifulsoup4==4.12.3
